Added scripts generating first test examples
Added two python scripts which generates files with examples of correct
lemmatization of the words which admit only one lemmatization regardless
of the context they are used in.

The scripts use the morphological dictionary and some frequency list of
the OpenCorpora project (http://opencorpora.org/), and the pymorphy2
package (https://pymorphy2.readthedocs.io/en/latest/) for morphological
analysis.

See more details in the scripts' comments.